Time: 0 ms (100%), Space: 8.8 MB (35.05%) - LeetHub

+++ 3379-score-of-a-string/3379-score-of-a-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int scoreOfString(string s) {
-        int result = 0;
+        unsigned short result = 0;
         unsigned char strLen = s.length();
 
         for (unsigned char i = 1; i < strLen; ++i) {
